requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
Pillow>=10.4.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.9.0
reportlab>=4.2.0
gunicorn==21.2.0
whitenoise==6.6.0
celery>=5.3.0
redis>=5.0.0
requests>=2.32.0
